+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the "git-changelists" extension will be documented in this file.
 
+### 0.1.6
+
+- Fix check if git repo exist
+- Add files counter to changelist
+- Add file open on click
+
 ### 0.1.5
 
 - Refactor error handling